DrawingTool: refactor types, add getAnswerType
